fix: 修复 GIF 导出失败 + 页面 TDZ 渲染崩溃
- gifExport: quantize/applyPalette 必须传逐字节 Uint8ClampedArray，
  原代码误传 Uint32Array 导致 gifenc 抛错使整个导出失败；
  d.ts 类型一并收紧为 Uint8Array | Uint8ClampedArray
- gifExport: delay 单位改毫秒（原传百分秒导致快 10 倍）；
  修正帧对齐偏移（多减了 PADDING）；显式设置透明索引避免闪色块
- App: 将 animPlayingRef 同步 useEffect 移到 animPlaying 声明之后，
  修复 Cannot access 'animPlaying' before initialization 致白屏

+++ src/types/gifenc.d.ts
@@ -46,14 +46,16 @@ declare module "gifenc" {
     clearAlpha?: boolean;
   }
 
+  // 注意：运行时必须是 flat 的 Uint8Array / Uint8ClampedArray（逐字节 RGBA），
+  // 传 Uint32Array 会让 gifenc 抛错。ImageData.data 是 Uint8ClampedArray，正好可用。
   export function quantize(
-    rgba: Uint8Array | Uint32Array,
+    rgba: Uint8Array | Uint8ClampedArray,
     maxColors: number,
     opts?: QuantizeOptions,
   ): number[][];
 
   export function applyPalette(
-    rgba: Uint8Array | Uint32Array,
+    rgba: Uint8Array | Uint8ClampedArray,
     palette: number[][],
     format?: string,
   ): Uint8Array;
